+++ src/wrappers/themis/wasm/src/context.ts
@@ -1,0 +1,6 @@
+
+import type {LibThemisModule} from "./libthemis";
+
+const context: { libthemis?: LibThemisModule } = {};
+
+export default context;